Mouse based movement and lasers!
Added movment based on the location of the mouse instead of just based
on the screen. When you click w the ship flies in the direction of the
mouse and when you click s it flies away. A and D will rotate the ship
around the mouse. When left click is pressed a laser is spawned and shot
in the direction of the cursor. It goes really slow currently.